+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Changes:
 * The deprecated flag `--gen-config conf` is removed
 * Flag `--gen-config` now does the same thing as `--migrate-config`, which can be used as config migration and default config file generation. Flag `--migrate-config` is removed
+* Fastfetch now searchs for config files in the order of `fastfetch --list-config-paths`, and won't load other config if one is found.
+* Flag `--load-user-config` now works in command line flags, but not in config file.
 
 # 2.2.3